Add when trait instead of code duplicate.
